CocoaInputMethod: Change commit method for composition
macos' composition is a little different to the other system.
Normally it does commit preedit not when the focus is changing,
but when the new input starts in another object.

This change makes the system-wide management for the composition
possible such as converting between Hiragana and Katakana in Japanese.

Fixes: QTBUG-115038
Task-number: QTBUG-59222
Pick-to: 6.7
Change-Id: Iff949707ec2fa1443c392509e491ffaaa0b5ab50